Event Handler and Dispatcher in action

+++ Nnuts/src/Nnuts/Events/MouseEvent.h
@@ -45,7 +45,7 @@ namespace Nnuts {
 
 	class NNUTS_API MouseButtonEvent : public Event {
 	public:
-		inline float GetMouseButton() const { return m_Button; }
+		inline int GetMouseButton() const { return m_Button; }
 		
 		EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
 	protected:
